+++ ui/js/store.js
@@ -2,6 +2,7 @@
 // Licensed under the MIT License
 
 const patchesURL = "/repos/patches"
+const packagesURL = "/repos/packages"
 
 async function getInstalledGameIds(){
     response = await fetch("/library/get");
@@ -19,6 +20,43 @@ async function installPatch(appId, patchIndex){
     });
 }
 
+async function fetchPackages(){
+    packages_response = await fetch(packagesURL);
+    packages = await packages_response.json();
+    packages = packages[0]
+    const packagesMenu = document.getElementById("options-0");
+    while (packagesMenu.firstChild) {
+        packagesMenu.firstChild.remove();
+    }
+    let packagesCount = 0;
+    
+    const firstSeparator = document.createElement("div");
+    firstSeparator.classList.add("option-separator");
+    packagesMenu.appendChild(firstSeparator);
+    
+    const reinstallingStr = translate_str("reinstalling...", "reinstalling...")
+    const installingStr = translate_str("installing...", "installing...")
+    const installStr = translate_str("install", "install");
+    const installedStr = translate_str("installed", "installed");
+    const noItemToListStr = translate_str("NoItemsToList", "No items to list");
+
+    if (packagesCount == 0) {
+        const emptyElement = document.createElement("div");
+        emptyElement.classList.add("menu-option");
+        emptyElement.classList.add("store-item");
+        emptyElement.innerText = noItemToListStr;
+        packagesMenu.appendChild(emptyElement);
+        const finalSeparator = document.createElement("div");
+        finalSeparator.classList.add("option-separator");
+        packagesMenu.appendChild(finalSeparator);
+    }
+
+    const spacer = document.createElement('div');
+    spacer.style.minHeight = "150px";
+    patchesMenu.appendChild(spacer);
+    refreshOptions();
+}
+
 async function fetchPatches(){
     patches_rp = await fetch(patchesURL);
     patches = await patches_rp.json();
@@ -33,6 +71,11 @@ async function fetchPatches(){
     const firstSeparator = document.createElement("div");
     firstSeparator.classList.add("option-separator");
     patchesMenu.appendChild(firstSeparator);
+    const reinstallingStr = translate_str("reinstalling...", "reinstalling...")
+    const installingStr = translate_str("installing...", "installing...")
+    const installStr = translate_str("install", "install");
+    const installedStr = translate_str("installed", "installed");
+    const noItemToListStr = translate_str("NoItemsToList", "No items to list");
 
     for (const [key, value] of Object.entries(patches)) {
         const AppID = key
@@ -58,7 +101,8 @@ async function fetchPatches(){
             patchElement.dataset.appid = AppID;
             patchElement.dataset.index = PatchIndex;
             patchElement.addEventListener("click", function(){
-                const installText = patchInstalled?"reinstalling...":"installing...";
+
+                const installText = patchInstalled?reinstallingStr:installingStr;
                 patchElement.getElementsByClassName("left-label")[0].innerText = installText;
                 installPatch(AppID, PatchIndex);
             });
@@ -66,9 +110,9 @@ async function fetchPatches(){
 
             const leftLabel = document.createElement("span");
             leftLabel.classList.add("left-label");
-            leftLabel.innerText = "install";
+            leftLabel.innerText = installStr;
             if (patchInstalled) {
-                leftLabel.innerText = "installed";
+                leftLabel.innerText = installedStr;
                 leftLabel.style.color = "orange";
             }
             patchElement.appendChild(leftLabel);
@@ -90,7 +134,7 @@ async function fetchPatches(){
         const emptyElement = document.createElement("div");
         emptyElement.classList.add("menu-option");
         emptyElement.classList.add("store-item");
-        emptyElement.innerText = "No items to list";
+        emptyElement.innerText = noItemToListStr;
         patchesMenu.appendChild(emptyElement);
         const finalSeparator = document.createElement("div");
         finalSeparator.classList.add("option-separator");